refactor: improve barrels, helpers & types
